changed the message layout and corrected some code
